UPDATE/REFACTORING - TileMaps Json for objects / Files refactor to convention
