+++ client/pages/login.js
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import Link from "next/link";
 import { login } from "../src/services/Login/ServicesLogin";
 import Alert from "../src/components/molecules/Alert/Alert";
-import { useUser } from "../src/context/userContext";
+import { useUser } from "../src/context/ContextUser";
 import { useRouter } from "next/router";
 import { getUserById } from "../src/services/User/ServicesGetUser";
 import Cookies from "js-cookie";
