+++ client/pages/register.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import SpritesCarousel from "../src/components/organisms/CarouselSprites/CarouselSprites";
 import { register } from "../src/services/Register/ServicesRegister";
-import { useUser } from "../src/context/userContext";
+import { useUser } from "../src/context/ContextUser";
 import { useRouter } from "next/router";
 import Link from "next/link";
 const SignUp = () => {
