+++ client/public/assets/tilemaps/SnowTown/SnowTown-Indoor-D.json
@@ -760,33 +760,6 @@
         }, 
         {
          "draworder":"topdown",
-         "id":30,
-         "name":"Audio",
-         "objects":[],
-         "opacity":1,
-         "properties":[
-                {
-                 "name":"audioName",
-                 "type":"string",
-                 "value":"Indoor"
-                }, 
-                {
-                 "name":"loop",
-                 "type":"bool",
-                 "value":true
-                }, 
-                {
-                 "name":"volume",
-                 "type":"float",
-                 "value":0.4
-                }],
-         "type":"objectgroup",
-         "visible":true,
-         "x":0,
-         "y":0
-        }, 
-        {
-         "draworder":"topdown",
          "id":15,
          "name":"Zone",
          "objects":[
@@ -899,23 +872,6 @@
         }, 
         {
          "draworder":"topdown",
-         "id":29,
-         "name":"CanCrossRun",
-         "objects":[],
-         "opacity":1,
-         "properties":[
-                {
-                 "name":"playerCantCrossRun",
-                 "type":"bool",
-                 "value":false
-                }],
-         "type":"objectgroup",
-         "visible":true,
-         "x":0,
-         "y":0
-        }, 
-        {
-         "draworder":"topdown",
          "id":17,
          "name":"Worlds",
          "objects":[
@@ -938,66 +894,42 @@
         }, 
         {
          "draworder":"topdown",
-         "id":20,
-         "name":"Weather",
-         "objects":[
-                {
-                 "class":"",
-                 "height":19,
-                 "id":4,
-                 "name":"Nothing",
-                 "rotation":0,
-                 "text":
-                    {
-                     "text":"Bonjour Monde",
-                     "wrap":true
-                    },
-                 "visible":true,
-                 "width":108,
-                 "x":464,
-                 "y":208
-                }],
-         "opacity":1,
-         "type":"objectgroup",
-         "visible":true,
-         "x":0,
-         "y":0
-        }, 
-        {
-         "draworder":"topdown",
-         "id":27,
-         "name":"MapName",
-         "objects":[
-                {
-                 "class":"",
-                 "height":19,
-                 "id":20,
-                 "name":"map-name",
-                 "rotation":0,
-                 "text":
-                    {
-                     "halign":"center",
-                     "text":"Maison porte D",
-                     "wrap":true
-                    },
-                 "visible":true,
-                 "width":130,
-                 "x":454,
-                 "y":410
-                }],
+         "id":31,
+         "name":"MapProperties",
+         "objects":[],
          "opacity":1,
          "properties":[
                 {
-                 "name":"name",
+                 "name":"audio",
                  "type":"string",
-                 "value":"Maison Porte E"
+                 "value":"Indoor|true|0.5|1.0"
+                }, 
+                {
+                 "name":"bike",
+                 "type":"bool",
+                 "value":false
+                }, 
+                {
+                 "name":"isNight",
+                 "type":"bool",
+                 "value":false
+                }, 
+                {
+                 "name":"mapName",
+                 "type":"string",
+                 "value":"Maison Porte D"
+                }, 
+                {
+                 "name":"weather",
+                 "type":"string",
+                 "value":"false|Default"
                 }],
          "type":"objectgroup",
          "visible":true,
          "x":0,
          "y":0
         }],
- "nextlayerid":31,
+ "nextlayerid":32,
  "nextobjectid":22,
  "orientation":"orthogonal",
  "properties":[
