+++ client/src/scenes/BootGame.js
@@ -128,27 +128,27 @@ export default class BootGame extends Scene {
     /** LOAD JSON SNOWTOWN DOOR A */
     this.load.tilemapTiledJSON(
       "SnowTown-DoorA",
-      "assets/tilemaps/SnowTown/SnowTown-DoorA.json"
+      "assets/tilemaps/SnowTown/SnowTown-Indoor-A.json"
     );
     /** LOAD JSON SNOWTOWN DOOR B */
     this.load.tilemapTiledJSON(
       "SnowTown-DoorB",
-      "assets/tilemaps/SnowTown/SnowTown-DoorB.json"
+      "assets/tilemaps/SnowTown/SnowTown-InDoor-B.json"
     );
     /** LOAD JSON SNOWTOWN DOOR C */
     this.load.tilemapTiledJSON(
       "SnowTown-DoorC",
-      "assets/tilemaps/SnowTown/SnowTown-DoorC.json"
+      "assets/tilemaps/SnowTown/SnowTown-Indoor-C.json"
     );
     /** LOAD JSON SNOWTOWN DOOR D */
     this.load.tilemapTiledJSON(
       "SnowTown-DoorD",
-      "assets/tilemaps/SnowTown/SnowTown-DoorD.json"
+      "assets/tilemaps/SnowTown/SnowTown-Indoor-D.json"
     );
     /** LOAD JSON SNOWTOWN DOOR E */
     this.load.tilemapTiledJSON(
       "SnowTown-DoorE",
-      "assets/tilemaps/SnowTown/SnowTown-DoorE.json"
+      "assets/tilemaps/SnowTown/SnowTown-Indoor-E.json"
     );
   }
 }
